Compatibility for ndarray batches of images
It was not possible to pass the mtcnn forward method a batch of n images stacked on the first dimension, even if described as possible in the documentation and working with the detection methods. In this way it is possible to pass a batch of np ndarrays images.
I have tested it against a batch of np ndarray images, a list of ndarrays images, a list of PIL images, single PIL and np ndarrays images.
Great package anyway!